Fix file system request init and cleanup
We did not use fs init and cleanup as uv 1.10.1 is.
The code has shown smaller footprint.
But it introduced memory leak problem when it is not
managed carefully. Thus we want to use uv's as it is.

In addition, I changed tuv_usleep to uv_sleep in test/raw/*.

libtuv-DCO-1.0-Signed-off-by: Sanggyu Lee sg5.lee@samsung.com