Report error if TARGET_PLATFORM was not defined
When calling the cmake for the project if
the TARGET_PLATFORM definition is not set only a cryptic
file not found error is displayed for the user.

By checking the TARGET_PLATFORM cmake definition
we can avoid the file not found error and display
a more descriptive reason why the cmake call failed.

libtuv-DCO-1.0-Signed-off-by: Peter Gal pgal.u-szeged@partner.samsung.com