Fix uv__to_stat function on Nuttx with newer gcc
With a newer arm gcc (5.0+) the _DEFAULT_SOURCE macro
gets defined when the unix/fs.c file is built thus
trying to access a set of stat struct members which
are not defined on Nuttx.

By adding an extra check to see if the build is for
Nuttx we can correctly select the stat conversion
code path.

libtuv-DCO-1.0-Signed-off-by: Peter Gal pgal.u-szeged@partner.samsung.com